moved helper functions to toss.helper
