+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "pactum",
-  "version": "3.0.3",
+  "version": "3.0.4",
   "description": "REST API Testing Tool for all levels of testing in a Test Pyramid",
   "main": "./src/index.js",
   "types": "./src/index.d.ts",
